+++ code/tests/cases/test_memory.cpp
@@ -411,7 +411,7 @@ FOSSIL_TEST(cpp_test_memory_class_stats) {
     fossil::sys::Memory::stats(&allocs, &bytes);
     // Just check that stats call does not crash and returns something
     ASSUME_ITS_TRUE(allocs >= (size_t)0);
-    ASSUME_ITS_TRUE(bytes >= 0);
+    ASSUME_ITS_TRUE(bytes >= (size_t)0);
 }
 
 // * * * * * * * * * * * * * * * * * * * * * * * *
